+++ src/components/button/button_empty/button_empty.tsx
@@ -51,7 +51,7 @@ const flushTypeToClassNameMap = {
 
 export const FLUSH_TYPES = keysOf(flushTypeToClassNameMap);
 
-export interface EuiButtonEmptyProps extends CommonProps {
+interface CommonEuiButtonEmptyProps extends CommonProps {
   iconType?: IconType;
   iconSide?: keyof typeof iconSideToClassNameMap;
   color?: EuiButtonEmptyColor;
@@ -80,16 +80,16 @@ export interface EuiButtonEmptyProps extends CommonProps {
   textProps?: Partial<HTMLAttributes<HTMLSpanElement>>;
 }
 
-type EuiButtonEmptyPropsForAnchor = PropsForAnchor<EuiButtonEmptyProps>;
+type EuiButtonEmptyPropsForAnchor = PropsForAnchor<CommonEuiButtonEmptyProps>;
 
-type EuiButtonEmptyPropsForButton = PropsForButton<EuiButtonEmptyProps>;
+type EuiButtonEmptyPropsForButton = PropsForButton<CommonEuiButtonEmptyProps>;
 
-type Props = ExclusiveUnion<
+export type EuiButtonEmptyProps = ExclusiveUnion<
   EuiButtonEmptyPropsForAnchor,
   EuiButtonEmptyPropsForButton
 >;
 
-export const EuiButtonEmpty: FunctionComponent<Props> = ({
+export const EuiButtonEmpty: FunctionComponent<EuiButtonEmptyProps> = ({
   children,
   className,
   iconType,